WebSocketのサンプルをモジュール化
