+++ backend/src/quiz/mod.rs
@@ -1,9 +1,8 @@
-
-use actix_web::{App, get, HttpResponse, HttpServer, middleware::Logger, Responder, web};
+use actix_web::{get, HttpResponse, Responder};
 
 mod quiz_json_reader;
 
-#[get("/mod")]
+#[get("/quiz")]
 pub async fn quiz() -> impl Responder {
     HttpResponse::Ok().json(quiz_json_reader::read_file())
 }
